+++ service/collector/cluster.go
@@ -4,14 +4,14 @@ import (
 	"context"
 	"fmt"
 
-	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
-	apiextensionsconditions "github.com/giantswarm/apiextensions/v3/pkg/conditions"
-	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v6/pkg/apis/infrastructure/v1alpha3"
+	apiextensionsconditions "github.com/giantswarm/apiextensions/v6/pkg/conditions"
+	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/prometheus/client_golang/prometheus"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
+	apiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/cluster-api/util/conditions"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -78,7 +78,7 @@ func NewCluster(config ClusterConfig) (*Cluster, error) {
 func (c *Cluster) Collect(ch chan<- prometheus.Metric) error {
 	ctx := context.Background()
 
-	var list apiv1alpha3.ClusterList
+	var list apiv1beta1.ClusterList
 	{
 		err := c.k8sClient.CtrlClient().List(
 			ctx,
@@ -165,7 +165,7 @@ func (c *Cluster) Collect(ch chan<- prometheus.Metric) error {
 			ch <- prometheus.MustNewConstMetric(
 				clusterStatus,
 				prometheus.GaugeValue,
-				boolToFloat64(conditions.IsTrue(&cl, apiv1alpha3.ReadyCondition)),
+				boolToFloat64(conditions.IsTrue(&cl, apiv1beta1.ReadyCondition)),
 				key.ClusterID(&cl),
 				key.ReleaseVersion(&cl),
 				infrastructurev1alpha3.ClusterStatusConditionCreated,