+++ service/collector/node_pool.go
@@ -3,11 +3,11 @@ package collector
 import (
 	"context"
 
-	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/prometheus/client_golang/prometheus"
-	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
+	apiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cluster-operator/v3/pkg/label"
@@ -75,7 +75,7 @@ func NewNodePool(config NodePoolConfig) (*NodePool, error) {
 func (np *NodePool) Collect(ch chan<- prometheus.Metric) error {
 	ctx := context.Background()
 
-	var list apiv1alpha3.MachineDeploymentList
+	var list apiv1beta1.MachineDeploymentList
 	{
 		err := np.k8sClient.CtrlClient().List(
 			ctx,