+++ service/internal/basedomain/internal/cache/cluster.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/giantswarm/operatorkit/v5/pkg/controller/context/cachekeycontext"
+	"github.com/giantswarm/operatorkit/v7/pkg/controller/context/cachekeycontext"
 	gocache "github.com/patrickmn/go-cache"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"