+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update ClusterAPI CR's to `v1beta1`.
+
 ## [3.14.1] - 2022-03-11
 
 ### Changed